Fix exc_stack GC by making exc stacks always belong to tasks
Make exc_stack always belong to a task and never to the TLS, to
correctly handle exception stacks as GC'd buffers.  A downside is that
we now allocate on the first throw within any given task. This should
generally be ok, but could loose the root cause of an error if we
coincidentally exhaust memory at the same time.

Also simplify save and restore of exception stack state in codegen and
interpreter by removing mention of this from jl_handler_t, and providing
a runtime function jl_exc_stack_state().