Fix tests for task switching inside finalizers
If these tests run in the interpreter rather than from compiled code,
the garbage is still reachable by the time we're trying to force
collection with GC.gc(). This is because the interpreter roots all
local variables until the end of the function call.

This small change ensures that the garbage never gets rooted in the
local variables of the main test code.

Previously the test code would not run in the interpreter, because of a
ccall in the lowering of `finally`.